sys-apps/pkgcraft-tools: add live ebuild
Using the official upstream ebuild as a base.

Compared to the live ebuild in the official gentoo repo, this uses a
shared library for scallop which should fix all the LTO issues filed for
pkgcraft-related packages since static libraries and LTO generally don't
work well together. This is done in a non-optional fashion since the
default Gentoo environment builds dynamically linked executables anyway.

In addition, unused dependencies are dropped since libgit2 support was
replaced with gitoxide a while ago.